+++ kinto/views/permissions.py
@@ -67,6 +67,29 @@ class PermissionsModel:
         include_deleted=False,
         parent_id=None,
     ):
+        objects, _ = self._get_objects(
+            filters=filters,
+            sorting=sorting,
+            pagination_rules=pagination_rules,
+            limit=limit,
+            include_deleted=include_deleted,
+            parent_id=parent_id,
+        )
+        return objects
+
+    def count_objects(self, filters=None, parent_id=None):
+        _, count = self._get_objects(filters=filters, parent_id=parent_id)
+        return count
+
+    def _get_objects(
+        self,
+        filters=None,
+        sorting=None,
+        pagination_rules=None,
+        limit=None,
+        include_deleted=False,
+        parent_id=None,
+    ):
         # Invert the permissions inheritance tree.
         perms_descending_tree = {}
         for on_resource, tree in PERMISSIONS_INHERITANCE_TREE.items():
@@ -98,7 +121,7 @@ class PermissionsModel:
         allowed_resources = {"bucket", "collection", "group"} & set(from_settings.keys())
         if allowed_resources:
             storage = self.request.registry.storage
-            every_bucket, _ = storage.get_all(parent_id="", resource_name="bucket")
+            every_bucket = storage.list_all(parent_id="", resource_name="bucket")
             for bucket in every_bucket:
                 bucket_uri = "/buckets/{id}".format_map(bucket)
                 for res in allowed_resources:
@@ -109,7 +132,7 @@ class PermissionsModel:
                         continue
                     # Fetch bucket collections and groups.
                     # XXX: wrong approach: query in a loop!
-                    every_subobjects, _ = storage.get_all(parent_id=bucket_uri, resource_name=res)
+                    every_subobjects = storage.list_all(parent_id=bucket_uri, resource_name=res)
                     for subobject in every_subobjects:
                         subobj_uri = bucket_uri + f"/{res}s/{subobject['id']}"
                         perms_by_object_uri.setdefault(subobj_uri, set()).update(resource_perms)
@@ -177,7 +200,7 @@ class PermissionsSchema(resource.ResourceSchema):
     description="List of user permissions",
     plural_path="/permissions",
     object_path=None,
-    collection_methods=("GET",),
+    plural_methods=("HEAD", "GET"),
 )
 class Permissions(resource.Resource):