+++ tests/core/test_storage.py
@@ -62,7 +62,8 @@ class StorageBaseTest(unittest.TestCase):
             (self.storage.delete, "", "", ""),
             (self.storage.delete_all, "", ""),
             (self.storage.purge_deleted, "", ""),
-            (self.storage.get_all, "", ""),
+            (self.storage.list_all, "", ""),
+            (self.storage.count_all, "", ""),
         ]
         for call in calls:
             self.assertRaises(NotImplementedError, *call)
@@ -137,7 +138,7 @@ class PostgreSQLStorageTest(StorageTest, unittest.TestCase):
         "storage_max_fetch_size": 10000,
         "storage_backend": "kinto.core.storage.postgresql",
         "storage_poolclass": "sqlalchemy.pool.StaticPool",
-        "storage_url": "postgres://postgres:postgres@localhost:5432/testdb",
+        "storage_url": "postgresql://postgres:postgres@localhost:5432/testdb",
         "storage_strict_json": True,
     }
 
@@ -151,16 +152,17 @@ class PostgreSQLStorageTest(StorageTest, unittest.TestCase):
         for i in range(4):
             self.create_object({"phone": "tel-{}".format(i)})
 
-        results, count = self.storage.get_all(**self.storage_kw)
+        results = self.storage.list_all(**self.storage_kw)
         self.assertEqual(len(results), 4)
 
         settings = {**self.settings, "storage_max_fetch_size": 2}
         config = self._get_config(settings=settings)
         limited = self.backend.load_from_config(config)
 
-        results, count = limited.get_all(**self.storage_kw)
-        self.assertEqual(count, 4)
+        results = limited.list_all(**self.storage_kw)
         self.assertEqual(len(results), 2)
+        count = limited.count_all(**self.storage_kw)
+        self.assertEqual(count, 4)
 
     def test_number_of_fetched_objects_is_per_page(self):
         for i in range(10):
@@ -170,11 +172,12 @@ class PostgreSQLStorageTest(StorageTest, unittest.TestCase):
         config = self._get_config(settings=settings)
         backend = self.backend.load_from_config(config)
 
-        results, count = backend.get_all(
+        results = backend.list_all(
             pagination_rules=[[Filter("number", 1, COMPARISON.GT)]], **self.storage_kw
         )
-        self.assertEqual(count, 10)
         self.assertEqual(len(results), 2)
+        count = backend.count_all(**self.storage_kw)
+        self.assertEqual(count, 10)
 
     def test_connection_is_rolledback_if_error_occurs(self):
         with self.storage.client.connect() as conn:
@@ -217,9 +220,9 @@ class PostgreSQLStorageTest(StorageTest, unittest.TestCase):
             self.backend.load_from_config(self._get_config(settings=settings))
             mocked.assert_any_call(msg)
 
-    def test_get_all_raises_if_missing_on_strange_query(self):
+    def test_list_all_raises_if_missing_on_strange_query(self):
         with self.assertRaises(ValueError):
-            self.storage.get_all(
+            self.storage.list_all(
                 "some-resource", "some-parent", filters=[Filter("author", MISSING, COMPARISON.HAS)]
             )
 
@@ -281,21 +284,21 @@ class PaginatedTest(unittest.TestCase):
         ]
 
         def sample_objects_side_effect(*args, **kwargs):
-            return (self.sample_objects, len(self.sample_objects))
+            return self.sample_objects
 
-        self.storage.get_all.side_effect = sample_objects_side_effect
+        self.storage.list_all.side_effect = sample_objects_side_effect
 
     def test_paginated_passes_sort(self):
         i = paginated(self.storage, sorting=[Sort("id", -1)])
         next(i)  # make the generator do anything
-        self.storage.get_all.assert_called_with(
+        self.storage.list_all.assert_called_with(
             sorting=[Sort("id", -1)], limit=25, pagination_rules=None
         )
 
     def test_paginated_passes_batch_size(self):
         i = paginated(self.storage, sorting=[Sort("id", -1)], batch_size=17)
         next(i)  # make the generator do anything
-        self.storage.get_all.assert_called_with(
+        self.storage.list_all.assert_called_with(
             sorting=[Sort("id", -1)], limit=17, pagination_rules=None
         )
 
@@ -307,15 +310,15 @@ class PaginatedTest(unittest.TestCase):
         objects = self.sample_objects
         objects.reverse()
 
-        def get_all_mock(*args, **kwargs):
+        def list_all_mock(*args, **kwargs):
             this_objects = objects[:3]
             del objects[:3]
-            return this_objects, len(this_objects)
+            return this_objects
 
-        self.storage.get_all.side_effect = get_all_mock
+        self.storage.list_all.side_effect = list_all_mock
 
         list(paginated(self.storage, sorting=[Sort("id", -1)]))
-        assert self.storage.get_all.call_args_list == [
+        assert self.storage.list_all.call_args_list == [
             mock.call(sorting=[Sort("id", -1)], limit=25, pagination_rules=None),
             mock.call(
                 sorting=[Sort("id", -1)],