+++ tests/plugins/test_quotas.py
@@ -72,7 +72,7 @@ class QuotaWebTest(support.BaseWebTest, unittest.TestCase):
 
         # Setup the postgresql backend for transaction support.
         settings["storage_backend"] = "kinto.core.storage.postgresql"
-        db = "postgres://postgres:postgres@localhost/testdb"
+        db = "postgresql://postgres:postgres@localhost/testdb"
         settings["storage_url"] = db
         settings["permission_backend"] = "kinto.core.permission.postgresql"
         settings["permission_url"] = db
@@ -138,7 +138,7 @@ class QuotaListenerTest(QuotaWebTest):
 
         self.app.delete("/buckets", headers=self.headers)
 
-        stored_in_backend, _ = self.storage.get_all(
+        stored_in_backend = self.storage.list_all(
             parent_id="/buckets/*", resource_name=QUOTA_RESOURCE_NAME
         )
         assert len(stored_in_backend) == 0
@@ -146,7 +146,7 @@ class QuotaListenerTest(QuotaWebTest):
     def test_bucket_delete_destroys_its_quota_entries(self):
         self.create_bucket()
         self.app.delete(self.bucket_uri, headers=self.headers)
-        stored_in_backend, _ = self.storage.get_all(
+        stored_in_backend = self.storage.list_all(
             parent_id="/buckets/test", resource_name=QUOTA_RESOURCE_NAME
         )
         assert len(stored_in_backend) == 0
@@ -237,7 +237,7 @@ class QuotaListenerTest(QuotaWebTest):
         self.create_bucket()
         self.create_collection()
         self.app.delete(self.collection_uri, headers=self.headers)
-        stored_in_backend, _ = self.storage.get_all(
+        stored_in_backend = self.storage.list_all(
             parent_id=self.collection_uri, resource_name=QUOTA_RESOURCE_NAME
         )
         assert len(stored_in_backend) == 0